+++ marple/src/dd/kms/marple/impl/gui/common/AccessModifierInput.java
@@ -25,7 +25,10 @@ public class AccessModifierInput extends JPanel
 		super(new GridBagLayout());
 
 		add(slider,					new GridBagConstraints(0, 0, 1, 1, 1.0, 0.0, WEST, HORIZONTAL, DEFAULT_INSETS, 0, 0));
-		add(accessModifierLabel,	new GridBagConstraints(0, 1, 1, 1, 1.0, 0.0, CENTER, HORIZONTAL, new Insets(3, 5, 10, 5), 0, 0));
+		add(accessModifierLabel,	new GridBagConstraints(1, 0, 1, 1, 0.0, 0.0, WEST, HORIZONTAL, DEFAULT_INSETS, 0, 0));
+
+		setPreferredSliderSize();
+		setPreferredAccessModifierLabelSize();
 
 		accessModifierLabel.setHorizontalAlignment(SwingConstants.CENTER);
 		accessModifierLabel.setOpaque(true);
@@ -37,6 +40,26 @@ public class AccessModifierInput extends JPanel
 
 	private void addListeners() {
 		slider.addChangeListener(e -> onSliderChanged());
+	}
+
+	private void setPreferredSliderSize() {
+		Dimension preferredSize = slider.getPreferredSize();
+		if (preferredSize == null) {
+			return;
+		}
+		slider.setPreferredSize(new Dimension(100, preferredSize.height));
+	}
+
+	private void setPreferredAccessModifierLabelSize() {
+		String oldText = accessModifierLabel.getText();
+		Dimension maxPreferredSize = new Dimension(0, 0);
+		for (AccessModifier accessModifier : AccessModifier.values()) {
+			accessModifierLabel.setText(accessModifier.toString());
+			Dimension preferredSize = accessModifierLabel.getPreferredSize();
+			maxPreferredSize = new Dimension(Math.max(maxPreferredSize.width, preferredSize.width + 20), Math.max(maxPreferredSize.height, preferredSize.height));
+		}
+		accessModifierLabel.setText(oldText);
+		accessModifierLabel.setPreferredSize(maxPreferredSize);
 	}
 
 	private void updateLabelText() {